refactor: revert some changes in the organization
